feat: implement UI for menu screen, add to MenuViewModel pagination chore: refactored database flow communication

+++ app/src/main/java/com/app/bitcoinwallet/domain/utils/WalletDataStore.kt
@@ -12,7 +12,7 @@ interface WalletDataStore {
 
     fun getBalance(): Flow<Float?>
 
-    suspend fun saveBitcoinRateUsd(rate: Float)
+    suspend fun saveBitcoinRateUsd(rate: Int)
 
-    fun getBitcoinRateUsd(): Flow<Float?>
+    fun getBitcoinRateUsd(): Flow<Int?>
 }
